refactor: Move shadcn directory to shared/ui
